// src/store/inventory/inventory.types.ts
export enum INVENTORY_ACTION_TYPES {
  FETCH_INVENTORY_START = 'FETCH_INVENTORY_START',
  FETCH_INVENTORY_SUCCESS = 'FETCH_INVENTORY_SUCCESS',
  FETCH_INVENTORY_FAILED = 'FETCH_INVENTORY_FAILED',

  POST_INVENTORY_START = 'POST_INVENTORY_START',
  POST_INVENTORY_SUCCESS = 'POST_INVENTORY_SUCCESS',
  POST_INVENTORY_FAILED = 'POST_INVENTORY_FAILED',

  PUT_INVENTORY_START = 'PUT_INVENTORY_START',
  PUT_INVENTORY_SUCCESS = 'PUT_INVENTORY_SUCCESS',
  PUT_INVENTORY_FAILED = 'PUT_INVENTORY_FAILED',

  DELETE_INVENTORY_START = 'DELETE_INVENTORY_START',
  DELETE_INVENTORY_SUCCESS = 'DELETE_INVENTORY_SUCCESS',
  DELETE_INVENTORY_FAILED = 'DELETE_INVENTORY_FAILED',
}

export type InventoryProps = {
  uid: string;
  title: string;
  createdAt: string;
  isActive: boolean;
};
export type InventoryState = {
  isLoading: boolean;
  error: Error | null;
  inventory: InventoryProps[] | [];
};